add merge for arbitrary dict, list nesting
