+++ saga/data_types/mixed_data_type/MixedDataType.py
@@ -1,27 +1,29 @@
-from saga.data_types.multi_dim_list.lcs import changed_paths, inserted_paths, removed_paths, lcs_multi_dimension, get_matching_path
-from saga.data_types.multi_dim_list.OP_MDL_Change import OP_MDL_Change
-from saga.data_types.multi_dim_list.OP_MDL_Insert import OP_MDL_Insert
-from saga.data_types.multi_dim_list.OP_MDL_Remove import OP_MDL_Remove
-from saga.data_types.multi_dim_list.mdl_merge_utils import diff3
-
 def value_at_path(obj, path):
     if len(path) == 0:
         return obj
     else:
         return value_at_path(obj[path[0]], path[1:])
 
+def dict_same_paths(dict_a, dict_b):
+    return {k for k in dict_a if k in dict_b and dict_a[k] == dict_b[k]}
+
+def dict_removed_paths(dict_a, dict_b):
+    return set(dict_a.keys()).difference(dict_b.keys())
+
+def dict_inserted_paths(dict_a, dict_b):
+    return removed_paths(dict_b, dict_a)
+
+def dict_changed_paths(dict_a, dict_b):
+    return {k for k in dict_a if k in dict_b and dict_a[k] != dict_b[k]}
+
 class MixedDataType(object):
 
     def __init__(self, mixed_data_type):
-        """
-        A mixed_data_type is a list or a dictionary. If the object is a list, it 
-        must contain only one type beneath it. If it is a dictonary, it can contain mulitple
-        types? 
-
-        """
         self.mixed_data_type = mixed_data_type
 
     def get_operations(self, mixed_data_type_obj):
+        return []
+        """
         A = self.multi_dim_list
         B = mixed_data_type_obj.mixed_data_type
 
@@ -51,15 +53,14 @@ class MixedDataType(object):
             operations.append(OP_MDL_Change("id", path, value_at_path(A, a_path), value_at_path(B, path)))
         
         return operations
+        """
 
     def merge(self, a_mdt, b_mdt):
-        
+        O = self.mixed_data_type
+        A = a_mdt.mixed_data_type
+        B = b_mdt.mixed_data_type
 
+        return MixedDataType(merge_rec(O, A, B))
 
-        merge_res = diff3(a_mdl.multi_dim_list, self.multi_dim_list, b_mdl.multi_dim_list, self.dimension)
-        if merge_res is None:
-            return None
-
-        return MultiDimList(merge_res, self.dimension)
